Safari-611.1.10/LayoutTests/accessibility/ios-simulator/non-contiguous-link-expected.txt

This is a link that would normally appear as multiple elements.
This test makes sure that only one element is returned for the link object, instead of many elements that are all links.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.description is 'AXLabel: This is a link that would normally appear as multiple elements'
PASS obj.isIgnored is false
PASS successfullyParsed is true

TEST COMPLETE
